requirements.txt
accelerate==1.2.1
alabaster==0.7.13
anyio==4.0.0
appdirs==1.4.4
asttokens==2.2.1
autopep8==2.0.4
Babel==2.13.1
backcall==0.2.0
blobfile==3.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
cobra==0.22.1
comm==0.1.4
contourpy==1.0.7
cplex==22.1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
depinfo==1.7.0
dgl==1.1.2+cu118
dglgo==0.0.2
diskcache==5.6.3
distro==1.9.0
docutils==0.20.1
et_xmlfile==2.0.0
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.16.1
fonttools==4.39.4
fsspec==2024.10.0
future==0.18.3
gensim==4.3.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.27.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==5.12.0
ipdb==0.13.13
ipykernel==6.25.1
ipython==8.14.0
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
jiter==0.10.0
joblib==1.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
lightning-utilities==0.9.0
littleutils==0.2.2
llvmlite==0.43.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==2.8.8
node2vec==0.4.6
numba==0.60.0
numpy==1.23.3
numpydoc==1.6.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
ogb==1.3.6
openai==1.90.0
openpyxl==3.1.5
optlang==1.7.0
outdated==0.2.2
packaging==23.1
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prompt-toolkit==3.0.38
protobuf==5.29.2
psutil==5.9.5
ptflops==0.7.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycryptodomex==3.21.0
pydantic==1.10.12
pyg-lib==0.2.0+pt113cu116
Pygments==2.15.1
pynndescent==0.5.13
pyparsing==3.0.9
python-dateutil==2.8.2
python-libsbml==5.19.0
pytorch_optimizer==2.11.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
rdkit-pypi==2022.9.5
regex==2024.11.6
requests==2.30.0
rich==13.5.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.4.5
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
six==1.16.0
sklearn==0.0.post5
smart-open==6.3.0
sniffio==1.3.0
snowballstemmer==2.2.0
some-package==0.1
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
stack-data==0.6.2
swiglpk==5.0.8
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.1.0
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.0.1
torch==2.8.0
torch-cluster==1.6.1+pt113cu116
torch-geometric==2.3.1
torch-scatter==2.1.1+pt113cu116
torch-sparse==0.6.17+pt113cu116
torch-spline-conv==1.2.2+pt113cu116
torchaudio==0.13.1+cu116
torchinfo==1.8.0
torchmetrics==1.2.0
torchvision==0.14.1+cu116
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
transformers==4.47.1
triton==3.4.0
typer==0.9.0
typing_extensions==4.14.0
tzdata==2023.3
umap-learn==0.5.7
urllib3==2.0.2
volcengine-python-sdk==4.0.1
wcwidth==0.2.6
zipp==3.15.0
